+++ R/sphunif-package.R
@@ -19,15 +19,15 @@
 #' novel datasets and gives the replicability for the data applications/
 #' simulations in García-Portugués et al. (2021)
 #' <doi:10.1007/978-3-030-69944-4_12>, García-Portugués et al. (2023)
-#' <doi:10.3150/21-BEJ1454>, García-Portugués et al. (2024)
-#' <doi:10.48550/arXiv.2108.09874>, and Fernández-de-Marcos and
-#' García-Portugués (2024) <doi:10.48550/arXiv.405.13531>.
+#' <doi:10.3150/21-BEJ1454>, Fernández-de-Marcos and García-Portugués (2024)
+#' <doi:10.1016/j.spl.2024.110218> and García-Portugués et al. (2024)
+#' <doi:10.48550/arXiv.2108.09874>.
 #'
 #' @author Eduardo García-Portugués and Thomas Verdebout.
 #' @references
 #' Fernández-de-Marcos, A. and García-Portugués, E. (2024) A stereographic test
-#' of spherical uniformity. \emph{arXiv:2405.13531}.
-#' \doi{10.48550/arXiv.2405.13531}.
+#' of spherical uniformity. \emph{Statistics & Probability Letters},
+#' 215:110218. \doi{doi:10.1016/j.spl.2024.110218}.
 #'
 #' García-Portugués, E. and Verdebout, T. (2018) An overview of uniformity
 #' tests on the hypersphere. \emph{arXiv:1804.00286}.